moved 'OkHttpClient.java' to package 'http'
