+++ recordreplay-integrationtest/src/test/java/be/xplore/recordreplay/IntegrationTestBase.java
@@ -4,7 +4,7 @@ import be.xplore.fakes.model.Stub;
 import be.xplore.fakes.service.DefaultHttpClient;
 import be.xplore.recordreplay.config.RecordReplayConfig;
 import be.xplore.recordreplay.junit4.RecordReplayRule;
-import be.xplore.recordreplay.service.OkHttpClient;
+import be.xplore.recordreplay.http.OkHttpClient;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.Before;
 import org.junit.Rule;
